third_party/blink/web_tests/fast/css-grid-layout/grid-auto-repeat-huge-grid-001.html

<!DOCTYPE html>
<title>Test for auto-fit and auto-fill with huge grids (lots of tracks)</title>
<link href="resources/grid.css" rel="stylesheet">
<link href="../css-intrinsic-dimensions/resources/width-keyword-classes.css" rel="stylesheet">
<link href="../css-intrinsic-dimensions/resources/height-keyword-classes.css" rel="stylesheet">
<link href="resources/huge-grids.css" rel="stylesheet">

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="resources/grid-definitions-parsing-utils.js"></script>
<script src="resources/grid-tracks-length.js"></script>

<div id="wideAutoFillGrid" class="grid wideGrid lotsOfFixedRepeatWithAutoFillCols"></div>
<div id="wideAutoFitGrid" class="grid wideGrid lotsOfFixedRepeatWithAutoFitCols">
    <div>Item1</div>
    <div>Item2</div>
    <div class="lastColumn">Item3</div>
</div>

<div id="autoFillGrid" class="grid tallGrid lotsOfFixedRepeatWithAutoFillRows"></div>
<div id="autoFitGrid" class="grid tallGrid lotsOfFixedRepeatWithAutoFitRows">
    <div>Item1</div>
    <div>Item2</div>
    <div class="lastRow">Item3</div>
</div>

<script>

test(function() {
     var autoFillGrid = testElement("wideAutoFillGrid", "grid-template-columns", 100000);
     var autoFitGrid = testElement("wideAutoFitGrid", "grid-template-columns", 100000);

     assert_equals(autoFillGrid[100000 - 1], "20px");
     assert_equals(autoFillGrid[100000 - 2], "7px");
     assert_equals(autoFillGrid[0], "10px");
     assert_equals(autoFillGrid[1], "2px");

     assert_equals(autoFitGrid[100000 - 1], "10px");
     assert_equals(autoFitGrid[100000 - 2], "0px");
     assert_equals(autoFitGrid[0], "10px");
     assert_equals(autoFitGrid[1], "2px");
}, "Test that we don't get more than kGridMaxTracks repetitions even on very wide grids (normal tracks clamped).");

test(function() {
     var autoFillGrid = testElement("autoFillGrid", "grid-template-rows", 100000);
     var autoFitGrid = testElement("autoFitGrid", "grid-template-rows", 100000);

     assert_equals(autoFillGrid[100000 - 1], "20px");
     assert_equals(autoFillGrid[100000 - 2], "7px");
     assert_equals(autoFillGrid[0], "10px");
     assert_equals(autoFillGrid[1], "2px");

     assert_equals(autoFitGrid[100000 - 1], "10px");
     assert_equals(autoFitGrid[100000 - 2], "0px");
     assert_equals(autoFitGrid[0], "10px");
     assert_equals(autoFitGrid[1], "2px");
}, "Test that we don't get more than kGridMaxTracks repetitions even on very tall grids (normal tracks clamped).");

</script>
